removing unneeded files
