+++ internal/controller/k8s_resources.go
@@ -7,7 +7,9 @@ import (
 	"encoding/json"
 	"errors"
 
+	pb "github.com/illumio/cloud-operator/api/illumio/cloud/k8scluster/v1"
 	"go.uber.org/zap"
+	"google.golang.org/protobuf/types/known/timestamppb"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -107,4 +109,24 @@ func getMetadatafromResource(logger *zap.SugaredLogger, resource unstructured.Un
 	} else {
 		return &metav1.ObjectMeta{}, errors.New("could not grab metadata from a resource")
 	}
+}
+
+// convertMetaObjectToMetadata takes a metav1.ObjectMeta and converts it into a proto message object KubernetesMetadata.
+func convertMetaObjectToMetadata(obj metav1.ObjectMeta, resource string) *pb.KubernetesObjectMetadata {
+	objMetadata := &pb.KubernetesObjectMetadata{
+		Annotations:       obj.GetAnnotations(),
+		CreationTimestamp: convertToProtoTimestamp(obj.CreationTimestamp),
+		Kind:              resource,
+		Labels:            obj.GetLabels(),
+		Name:              obj.GetName(),
+		Namespace:         obj.GetNamespace(),
+		ResourceVersion:   obj.GetResourceVersion(),
+		Uid:               string(obj.GetUID()),
+	}
+	return objMetadata
+}
+
+// convertToProtoTimestamp converts a Kubernetes metav1.Time into a Protobuf Timestamp.
+func convertToProtoTimestamp(k8sTime metav1.Time) *timestamppb.Timestamp {
+	return timestamppb.New(k8sTime.Time)
 }
