+++ internal/controller/k8s_resources_test.go
@@ -8,7 +8,10 @@ import (
 	"testing"
 	"time"
 
+	pb "github.com/illumio/cloud-operator/api/illumio/cloud/k8scluster/v1"
+	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
+	"google.golang.org/protobuf/types/known/timestamppb"
 	corev1 "k8s.io/api/core/v1"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -169,4 +172,41 @@ func TestGetMetadataFromResource(t *testing.T) {
 	if metadata.Name != "test-pod" || metadata.Namespace != "test-namespace" || metadata.Labels["app"] != "test-app" {
 		t.Errorf("Incorrect metadata extracted: %+v", metadata)
 	}
+}
+
+func TestConvertMetaObjectToMetadata(t *testing.T) {
+	sampleData := make(map[string]string)
+	resource := "test-resource"
+	creationTimestamp := metav1.Time{Time: time.Now()}
+	objMeta := metav1.ObjectMeta{
+		Annotations:       sampleData,
+		CreationTimestamp: creationTimestamp,
+		Labels:            sampleData,
+		Name:              "test-name",
+		Namespace:         "test-namespace",
+		ResourceVersion:   "test-version",
+		UID:               "test-uid",
+	}
+
+	expected := &pb.KubernetesObjectMetadata{
+		Annotations:       sampleData,
+		CreationTimestamp: convertToProtoTimestamp(creationTimestamp),
+		Kind:              resource,
+		Labels:            sampleData,
+		Name:              "test-name",
+		Namespace:         "test-namespace",
+		ResourceVersion:   "test-version",
+		Uid:               "test-uid",
+	}
+
+	result := convertMetaObjectToMetadata(objMeta, resource)
+	assert.Equal(t, expected, result)
+}
+
+func TestConvertToProtoTimestamp(t *testing.T) {
+	k8sTime := metav1.Time{Time: time.Now()}
+	expected := timestamppb.New(k8sTime.Time)
+
+	result := convertToProtoTimestamp(k8sTime)
+	assert.Equal(t, expected, result)
 }
