+++ internal/controller/secrets_test.go
@@ -5,6 +5,7 @@ package controller
 import (
 	"context"
 	"os"
+	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
@@ -232,4 +233,19 @@ func (suite *ControllerTestSuite) TestWriteK8sSecret() {
 			}
 		})
 	}
+}
+
+func TestIsRunningInCluster(t *testing.T) {
+	t.Run("Running in cluster", func(t *testing.T) {
+		os.Setenv("KUBERNETES_SERVICE_HOST", "localhost")
+		defer os.Unsetenv("KUBERNETES_SERVICE_HOST")
+
+		assert.True(t, IsRunningInCluster())
+	})
+
+	t.Run("Not running in cluster", func(t *testing.T) {
+		os.Unsetenv("KUBERNETES_SERVICE_HOST")
+
+		assert.False(t, IsRunningInCluster())
+	})
 }
