+++ internal/controller/testhelper/set_up_test_cluster.go
@@ -3,12 +3,14 @@
 package testhelper
 
 import (
+	"fmt"
 	"os/exec"
 )
 
 // setupTestCluster creates a new KIND cluster for testing.
 func SetupTestCluster() error {
 	cmd := exec.Command("kind", "create", "cluster", "--name", "my-test-cluster", "--config", "../../kind-config.yaml")
+	fmt.Println(cmd)
 	return cmd.Run()
 }
